backend/requirements.txt
fastapi
uvicorn
pydantic
pytest-playwright